Adding input and output folders for general, login, valid accounts file list, and create account
Created all of the input and output files needed for each of the tests for general, login, valid accounts files list, and create account requirements. These are R1, R2, R3, and R4, respectively. Also created the alternate valid file lists files needed for testing the constraints of the valid accounts files list. As well as creating the alternate valid file lists file for the create account constraint where you cannot create an account with the same number as an already valid account.